split runner file: remove unnecessary includes
