+++ sources/Runner/Runner.cpp
@@ -11,16 +11,8 @@
 #include <TraceLog.hpp>
 #include "System/Keyboard/KeyboardSystem.hpp"
 #include "System/Controllable/ControllableSystem.hpp"
-#include "Component/Movable/MovableComponent.hpp"
-#include "Component/Controllable/ControllableComponent.hpp"
-#include "Component/Keyboard/KeyboardComponent.hpp"
 #include "System/Gamepad/GamepadSystem.hpp"
 #include <System/Collision/CollisionSystem.hpp>
-#include "Component/Button/ButtonComponent.hpp"
-#include <Component/Collision/CollisionComponent.hpp>
-#include "Component/Renderer/CameraComponent.hpp"
-#include "Component/Renderer/Drawable3DComponent.hpp"
-#include "Component/Renderer/Drawable2DComponent.hpp"
 #include "Runner.hpp"
 #include "Models/GameState.hpp"
 #include <Model/ModelAnimations.hpp>
@@ -29,22 +21,14 @@
 #include <System/Event/EventSystem.hpp>
 #include <System/Health/HealthSystem.hpp>
 #include <System/Animator/AnimatorSystem.hpp>
-#include <Component/Animator/AnimatorComponent.hpp>
-#include <Component/IntroAnimation/IntroAnimationComponent.hpp>
 #include <System/IntroAnimation/IntroAnimationSystem.hpp>
 #include <System/Levitate/LevitateSystem.hpp>
 #include <System/Bonus/PlayerBonusSystem.hpp>
-#include <Component/Bonus/PlayerBonusComponent.hpp>
-#include <Component/Tag/TagComponent.hpp>
-#include "Component/Animation/AnimationsComponent.hpp"
 #include "System/Animation/AnimationsSystem.hpp"
-#include "Component/Shaders/ShaderComponent.hpp"
 #include "Map/Map.hpp"
 #include "System/MenuControllable/MenuControllableSystem.hpp"
 #include <Drawables/Texture.hpp>
 #include <System/Bomb/BombSystem.hpp>
-#include "Component/Music/MusicComponent.hpp"
-#include "Component/Sound/SoundComponent.hpp"
 #include "System/Sound/PlayerSoundManagerSystem.hpp"
 #include "System/Sound/MenuSoundManagerSystem.hpp"
 #include "System/Music/MusicSystem.hpp"
@@ -106,383 +90,6 @@ namespace BBM
 		wal.addSystem<AnimationsSystem>()
 			.addSystem<AnimatorSystem>()
 			.addSystem<RenderSystem>(window);
-	}
-
-	std::shared_ptr<WAL::Scene> Runner::loadMainMenuScene()
-	{
-		static const std::map<SoundComponent::SoundIndex, std::string> sounds = {
-			{SoundComponent::JUMP, "assets/sounds/click.ogg"}
-		};
-		auto scene = std::make_shared<WAL::Scene>();
-
-		scene->addEntity("Control entity")
-			.addComponent<ControllableComponent>()
-			.addComponent<KeyboardComponent>()
-			.addComponent<MusicComponent>("assets/musics/music_title.ogg")
-			.addComponent<SoundComponent>(sounds);
-		scene->addEntity("background")
-			.addComponent<PositionComponent>()
-			.addComponent<Drawable2DComponent, RAY::Texture>("assets/plain_menu_background.png");
-		scene->addEntity("logo")
-			.addComponent<PositionComponent>(1920 / 3, 180, 0)
-			.addComponent<Drawable2DComponent, RAY::Texture>("assets/logo_small.png");
-		auto &play = scene->addEntity("play button")
-			.addComponent<PositionComponent>(1920 / 2.5, 1080 - 540, 0)
-			.addComponent<Drawable2DComponent, RAY::Texture>("assets/buttons/button_new_game.png")
-			.addComponent<OnIdleComponent>([](WAL::Entity &entity, WAL::Wal &)
-			{
-				RAY::Texture *texture = dynamic_cast<RAY::Texture *>(entity.getComponent<Drawable2DComponent>().drawable.get());
-
-				texture->use("assets/buttons/button_new_game.png");
-			})
-			.addComponent<OnHoverComponent>([](WAL::Entity &entity, WAL::Wal &)
-			{
-				RAY::Texture *texture = dynamic_cast<RAY::Texture *>(entity.getComponent<Drawable2DComponent>().drawable.get());
-
-				texture->use("assets/buttons/button_new_game_hovered.png");
-			})
-			.addComponent<OnClickComponent>([](WAL::Entity &entity, WAL::Wal &)
-			{
-				gameState.nextScene = BBM::GameState::SceneID::GameScene;
-			});
-		auto &settings = scene->addEntity("settings button")
-			.addComponent<PositionComponent>(1920 / 2.5, 1080 - 360, 0)
-			.addComponent<Drawable2DComponent, RAY::Texture>("assets/buttons/button_settings.png")
-			.addComponent<OnIdleComponent>([](WAL::Entity &entity, WAL::Wal &)
-			{
-				RAY::Texture *texture = dynamic_cast<RAY::Texture *>(entity.getComponent<Drawable2DComponent>().drawable.get());
-
-				texture->use("assets/buttons/button_settings.png");
-			})
-			.addComponent<OnHoverComponent>([](WAL::Entity &entity, WAL::Wal &)
-			{
-				RAY::Texture *texture = dynamic_cast<RAY::Texture *>(entity.getComponent<Drawable2DComponent>().drawable.get());
-
-				texture->use("assets/buttons/button_settings_hovered.png");
-			})
-			.addComponent<OnClickComponent>([](WAL::Entity &entity, WAL::Wal &)
-			{
-				gameState.nextScene = BBM::GameState::SceneID::SettingsScene;
-			});
-		auto &exit = scene->addEntity("exit button")
-			.addComponent<PositionComponent>(1920 / 2.5, 1080 - 180, 0)
-			.addComponent<Drawable2DComponent, RAY::Texture>("assets/buttons/button_exit.png")
-			.addComponent<OnIdleComponent>([](WAL::Entity &entity, WAL::Wal &)
-			{
-				RAY::Texture *texture = dynamic_cast<RAY::Texture *>(entity.getComponent<Drawable2DComponent>().drawable.get());
-
-				texture->use("assets/buttons/button_exit.png");
-			})
-			.addComponent<OnHoverComponent>([](WAL::Entity &entity, WAL::Wal &)
-			{
-				RAY::Texture *texture = dynamic_cast<RAY::Texture *>(entity.getComponent<Drawable2DComponent>().drawable.get());
-
-				texture->use("assets/buttons/button_exit_hovered.png");
-			})
-			.addComponent<OnClickComponent>([](WAL::Entity &entity, WAL::Wal &wal)
-			{
-				wal.shouldClose = true;
-			});
-		auto &credits = scene->addEntity("credit button")
-			.addComponent<PositionComponent>(1920 - 100, 1080 - 30, 0)
-			.addComponent<Drawable2DComponent, RAY2D::Text>("Credits", 20, RAY::Vector2(), BLACK)
-			.addComponent<OnIdleComponent>([](WAL::Entity &entity, WAL::Wal &)
-			{
-				RAY2D::Text *text = dynamic_cast<RAY2D::Text *>(entity.getComponent<Drawable2DComponent>().drawable.get());
-
-				text->setColor(BLACK);
-			})
-			.addComponent<OnHoverComponent>([](WAL::Entity &entity, WAL::Wal &)
-			{
-				RAY2D::Text *text = dynamic_cast<RAY2D::Text *>(entity.getComponent<Drawable2DComponent>().drawable.get());
-
-				text->setColor(ORANGE);
-			})
-			.addComponent<OnClickComponent>([](WAL::Entity &entity, WAL::Wal &wal)
-			{
-				gameState.nextScene = BBM::GameState::SceneID::CreditScene;
-			});
-		play.getComponent<OnClickComponent>().setButtonLinks(nullptr, &settings);
-		settings.getComponent<OnClickComponent>().setButtonLinks(&play, &exit);
-		exit.getComponent<OnClickComponent>().setButtonLinks(&settings, &credits, nullptr, &credits);
-		credits.getComponent<OnClickComponent>().setButtonLinks(&exit, nullptr, &exit);
-		return scene;
-	}
-
-	std::shared_ptr<WAL::Scene> Runner::loadPauseMenuScene()
-	{
-		static const std::map<SoundComponent::SoundIndex, std::string> sounds = {
-			{SoundComponent::JUMP, "assets/sounds/click.ogg"}
-		};
-		auto scene = std::make_shared<WAL::Scene>();
-
-		scene->addEntity("Control entity")
-			.addComponent<ControllableComponent>()
-			.addComponent<KeyboardComponent>()
-			.addComponent<MusicComponent>("assets/musics/music_player_select.ogg")
-			.addComponent<SoundComponent>(sounds);
-		scene->addEntity("background")
-			.addComponent<PositionComponent>()
-			.addComponent<Drawable2DComponent, RAY::Texture>("assets/plain_menu_background.png");
-		scene->addEntity("pause text")
-			.addComponent<PositionComponent>(1920 / 2.5, 180, 0)
-			.addComponent<Drawable2DComponent, RAY2D::Text>("PAUSE", 120, RAY::Vector2(), ORANGE);
-		auto &play = scene->addEntity("play button")
-			.addComponent<PositionComponent>(1920 / 6.5, 1080 - 360, 0)
-			.addComponent<Drawable2DComponent, RAY::Texture>("assets/buttons/button_back.png")
-			.addComponent<OnIdleComponent>([](WAL::Entity &entity, WAL::Wal &)
-			{
-				RAY::Texture *texture = dynamic_cast<RAY::Texture *>(entity.getComponent<Drawable2DComponent>().drawable.get());
-
-				texture->use("assets/buttons/button_back.png");
-			})
-			.addComponent<OnHoverComponent>([](WAL::Entity &entity, WAL::Wal &)
-			{
-				RAY::Texture *texture = dynamic_cast<RAY::Texture *>(entity.getComponent<Drawable2DComponent>().drawable.get());
-
-				texture->use("assets/buttons/button_back_hovered.png");
-			})
-			.addComponent<OnClickComponent>([](WAL::Entity &entity, WAL::Wal &)
-			{
-				gameState.nextScene = BBM::GameState::SceneID::GameScene;
-			});
-		auto &settings = scene->addEntity("settings button")
-			.addComponent<PositionComponent>(1920 / 2.5, 1080 - 360, 0)
-			.addComponent<Drawable2DComponent, RAY::Texture>("assets/buttons/button_settings.png")
-			.addComponent<OnIdleComponent>([](WAL::Entity &entity, WAL::Wal &)
-			{
-				RAY::Texture *texture = dynamic_cast<RAY::Texture *>(entity.getComponent<Drawable2DComponent>().drawable.get());
-
-				texture->use("assets/buttons/button_settings.png");
-			})
-			.addComponent<OnHoverComponent>([](WAL::Entity &entity, WAL::Wal &)
-			{
-				RAY::Texture *texture = dynamic_cast<RAY::Texture *>(entity.getComponent<Drawable2DComponent>().drawable.get());
-
-				texture->use("assets/buttons/button_settings_hovered.png");
-			})
-			.addComponent<OnClickComponent>([](WAL::Entity &entity, WAL::Wal &)
-			{
-				gameState.nextScene = BBM::GameState::SceneID::SettingsScene;
-			});
-		auto &exit = scene->addEntity("exit button")
-			.addComponent<PositionComponent>(1920 / 1.5, 1080 - 360, 0)
-			.addComponent<Drawable2DComponent, RAY::Texture>("assets/buttons/button_exit.png")
-			.addComponent<OnIdleComponent>([](WAL::Entity &entity, WAL::Wal &)
-			{
-				RAY::Texture *texture = dynamic_cast<RAY::Texture *>(entity.getComponent<Drawable2DComponent>().drawable.get());
-
-				texture->use("assets/buttons/button_exit.png");
-			})
-			.addComponent<OnHoverComponent>([](WAL::Entity &entity, WAL::Wal &)
-			{
-				RAY::Texture *texture = dynamic_cast<RAY::Texture *>(entity.getComponent<Drawable2DComponent>().drawable.get());
-
-				texture->use("assets/buttons/button_exit_hovered.png");
-			})
-			.addComponent<OnClickComponent>([](WAL::Entity &entity, WAL::Wal &wal)
-			{
-				gameState.nextScene = BBM::GameState::SceneID::MainMenuScene;
-			});
-		//needed material
-		//music
-		play.getComponent<OnClickComponent>().setButtonLinks(nullptr, nullptr, nullptr, &settings);
-		settings.getComponent<OnClickComponent>().setButtonLinks(nullptr, nullptr, &play, &exit);
-		exit.getComponent<OnClickComponent>().setButtonLinks(nullptr, nullptr, &settings, nullptr);
-		return scene;
-	}
-
-	std::shared_ptr<WAL::Scene> Runner::loadSettingsMenuScene()
-	{
-		auto scene = std::make_shared<WAL::Scene>();
-		static const std::map<SoundComponent::SoundIndex, std::string> sounds = {
-			{SoundComponent::JUMP, "assets/sounds/click.ogg"}
-		};
-
-		scene->addEntity("Control entity")
-			.addComponent<ControllableComponent>()
-			.addComponent<KeyboardComponent>()
-			.addComponent<MusicComponent>("assets/musics/music_title.ogg")
-			.addComponent<SoundComponent>(sounds);
-		scene->addEntity("background")
-			.addComponent<PositionComponent>()
-			.addComponent<Drawable2DComponent, RAY::Texture>("assets/plain_menu_background.png");
-		scene->addEntity("logo")
-			.addComponent<PositionComponent>(1920 / 3, 180, 0)
-			.addComponent<Drawable2DComponent, RAY::Texture>("assets/logo_small.png");
-		auto &music = scene->addEntity("music text")
-			.addComponent<PositionComponent>(1920 / 2.5, 1080 - 540, 0)
-			.addComponent<Drawable2DComponent, RAY2D::Text>("Music Volume", 70, RAY::Vector2(), BLACK)
-			.addComponent<OnClickComponent>()
-			.addComponent<OnIdleComponent>([](WAL::Entity &entity, WAL::Wal &)
-			{
-				entity.getComponent<Drawable2DComponent>().drawable->setColor(BLACK);
-			})
-			.addComponent<OnHoverComponent>([](WAL::Entity &entity, WAL::Wal &)
-			{
-				entity.getComponent<Drawable2DComponent>().drawable->setColor(ORANGE);
-			});
-
-		auto &musicUp = scene->addEntity("music up button")
-			.addComponent<PositionComponent>(1920 / 1.5, 1080 - 540, 0)
-			.addComponent<Drawable2DComponent, RAY::Texture>("assets/buttons/button_plus.png")
-			.addComponent<MusicComponent>("assets/musics/music_title.ogg")
-			.addComponent<OnIdleComponent>([](WAL::Entity &entity, WAL::Wal &)
-			{
-				RAY::Texture *texture = dynamic_cast<RAY::Texture *>(entity.getComponent<Drawable2DComponent>().drawable.get());
-
-				texture->use("assets/buttons/button_plus.png");
-			})
-			.addComponent<OnClickComponent>([](WAL::Entity &entity, WAL::Wal &)
-			{
-				auto &component = entity.getComponent<MusicComponent>();
-
-				component.turnUpVolume();
-			})
-			.addComponent<OnHoverComponent>([](WAL::Entity &entity, WAL::Wal &)
-			{
-				RAY::Texture *texture = dynamic_cast<RAY::Texture *>(entity.getComponent<Drawable2DComponent>().drawable.get());
-
-				texture->use("assets/buttons/button_plus_hovered.png");
-			});
-
-		auto &musicDown = scene->addEntity("music down button")
-			.addComponent<PositionComponent>(1920 / 3, 1080 - 540, 0)
-			.addComponent<Drawable2DComponent, RAY::Texture>("assets/buttons/button_minus.png")
-			.addComponent<MusicComponent>("assets/musics/music_title.ogg")
-			.addComponent<OnIdleComponent>([](WAL::Entity &entity, WAL::Wal &)
-			{
-				RAY::Texture *texture = dynamic_cast<RAY::Texture *>(entity.getComponent<Drawable2DComponent>().drawable.get());
-
-				texture->use("assets/buttons/button_minus.png");
-			})
-			.addComponent<OnClickComponent>([](WAL::Entity &entity, WAL::Wal &)
-			{
-				auto &component = entity.getComponent<MusicComponent>();
-
-				component.turnDownVolume();
-			})
-			.addComponent<OnHoverComponent>([](WAL::Entity &entity, WAL::Wal &)
-			{
-				RAY::Texture *texture = dynamic_cast<RAY::Texture *>(entity.getComponent<Drawable2DComponent>().drawable.get());
-
-				texture->use("assets/buttons/button_minus_hovered.png");
-			});
-
-		auto &sound = scene->addEntity("sound text")
-			.addComponent<PositionComponent>(1920 / 2.5, 1080 - 360, 0)
-			.addComponent<Drawable2DComponent, RAY2D::Text>("Sound Volume", 70, RAY::Vector2(), BLACK)
-			.addComponent<OnClickComponent>()
-			.addComponent<OnIdleComponent>([](WAL::Entity &entity, WAL::Wal &)
-			{
-				entity.getComponent<Drawable2DComponent>().drawable->setColor(BLACK);
-			})
-			.addComponent<OnHoverComponent>([](WAL::Entity &entity, WAL::Wal &)
-			{
-				entity.getComponent<Drawable2DComponent>().drawable->setColor(ORANGE);
-			});
-
-		auto &soundUp = scene->addEntity("sound up button")
-			.addComponent<PositionComponent>(1920 / 1.5, 1080 - 360, 0)
-			.addComponent<Drawable2DComponent, RAY::Texture>("assets/buttons/button_plus.png")
-			.addComponent<SoundComponent>(sounds)
-			.addComponent<OnClickComponent>([](WAL::Entity &entity, WAL::Wal &)
-			{
-				auto &component = entity.getComponent<SoundComponent>();
-
-				component.turnUpVolume();
-			})
-			.addComponent<OnIdleComponent>([](WAL::Entity &entity, WAL::Wal &)
-			{
-				RAY::Texture *texture = dynamic_cast<RAY::Texture *>(entity.getComponent<Drawable2DComponent>().drawable.get());
-
-				texture->use("assets/buttons/button_plus.png");
-			})
-			.addComponent<OnHoverComponent>([](WAL::Entity &entity, WAL::Wal &)
-			{
-				RAY::Texture *texture = dynamic_cast<RAY::Texture *>(entity.getComponent<Drawable2DComponent>().drawable.get());
-
-				texture->use("assets/buttons/button_plus_hovered.png");
-			});
-
-		auto &soundDown = scene->addEntity("sound down button")
-			.addComponent<PositionComponent>(1920 / 3, 1080 - 360, 0)
-			.addComponent<Drawable2DComponent, RAY::Texture>("assets/buttons/button_minus.png")
-			.addComponent<SoundComponent>(sounds)
-			.addComponent<OnIdleComponent>([](WAL::Entity &entity, WAL::Wal &)
-			{
-				RAY::Texture *texture = dynamic_cast<RAY::Texture *>(entity.getComponent<Drawable2DComponent>().drawable.get());
-
-				texture->use("assets/buttons/button_minus.png");
-			})
-			.addComponent<OnClickComponent>([](WAL::Entity &entity, WAL::Wal &)
-			{
-				auto &component = entity.getComponent<SoundComponent>();
-
-				component.turnDownVolume();
-			})
-			.addComponent<OnHoverComponent>([](WAL::Entity &entity, WAL::Wal &)
-			{
-				RAY::Texture *texture = dynamic_cast<RAY::Texture *>(entity.getComponent<Drawable2DComponent>().drawable.get());
-
-				texture->use("assets/buttons/button_minus_hovered.png");
-			});
-
-		auto &debug = scene->addEntity("debug text")
-			.addComponent<PositionComponent>(1920 / 2.5, 1080 - 180, 0)
-			.addComponent<Drawable2DComponent, RAY2D::Text>("Debug Mode: Off", 70, RAY::Vector2(), BLACK)
-			.addComponent<OnClickComponent>([](WAL::Entity &entity, WAL::Wal &wal)
-			{
-				RAY2D::Text *text = dynamic_cast<RAY2D::Text *>(entity.getComponent<Drawable2DComponent>().drawable.get());
-
-				if (text->getString().find("Off") != std::string::npos) {
-					text->setText("Debug Mode: On");
-					wal.getSystem<RenderSystem>().setDebug(true);
-				} else {
-					text->setText("Debug Mode: Off");
-					wal.getSystem<RenderSystem>().setDebug(false);
-				}
-			})
-			.addComponent<OnIdleComponent>([](WAL::Entity &entity, WAL::Wal &)
-			{
-				entity.getComponent<Drawable2DComponent>().drawable->setColor(BLACK);
-			})
-			.addComponent<OnHoverComponent>([](WAL::Entity &entity, WAL::Wal &)
-			{
-				entity.getComponent<Drawable2DComponent>().drawable->setColor(ORANGE);
-			});
-		auto &back = scene->addEntity("back to menu")
-			.addComponent<PositionComponent>(10, 1080 - 85, 0)
-			.addComponent<Drawable2DComponent, RAY::Texture>("assets/buttons/button_back.png")
-			.addComponent<OnClickComponent>([](WAL::Entity &entity, WAL::Wal &)
-			{
-				gameState.nextScene = BBM::GameState::SceneID::MainMenuScene;
-			})
-			.addComponent<OnIdleComponent>([](WAL::Entity &entity, WAL::Wal &)
-			{
-				RAY::Texture *texture = dynamic_cast<RAY::Texture *>(entity.getComponent<Drawable2DComponent>().drawable.get());
-
-				texture->use("assets/buttons/button_back.png");
-			})
-			.addComponent<OnHoverComponent>([](WAL::Entity &entity, WAL::Wal &)
-			{
-				RAY::Texture *texture = dynamic_cast<RAY::Texture *>(entity.getComponent<Drawable2DComponent>().drawable.get());
-
-				texture->use("assets/buttons/button_back_hovered.png");
-			});
-		//needed material
-		//music
-		//sound
-
-		music.getComponent<OnClickComponent>().setButtonLinks(nullptr, &sound, &musicDown, &musicUp);
-		musicDown.getComponent<OnClickComponent>().setButtonLinks(&debug, &sound, nullptr, &music);
-		musicUp.getComponent<OnClickComponent>().setButtonLinks(&debug, &sound, &music);
-		sound.getComponent<OnClickComponent>().setButtonLinks(&music, &debug, &soundDown, &soundUp);
-		soundDown.getComponent<OnClickComponent>().setButtonLinks(&music, &debug, nullptr, &sound);
-		soundUp.getComponent<OnClickComponent>().setButtonLinks(&music, &debug, &sound);
-		debug.getComponent<OnClickComponent>().setButtonLinks(&sound, &back, &back);
-		back.getComponent<OnClickComponent>().setButtonLinks(&debug, nullptr, nullptr, &debug);
-		return scene;
 	}
 
 	void Runner::loadScenes()
